Implementation of  in the protocol
Added the use of  to handle warnings about unused variables.
Additionally, implemented background communication in the protocol, enhancing the interaction
and efficiency of data transmission between the client and server.
.

+++ client/clientProtocol.cpp
@@ -64,7 +64,7 @@ dataMatch ClientProtocol::ReceiveMatch(bool& isConnected) {
     }
 
     uint8_t newMatchID = assistant.reciveInt(wasClosed);
-    match.ID = newMatchID;
+    match.ID = static_cast<size_t>(newMatchID);
     uint8_t newQuantityPlayers = assistant.reciveInt(wasClosed);
     match.quantityPlayers = newQuantityPlayers;
     uint8_t newMAXPlayers = assistant.reciveInt(wasClosed);
