+++ server/serverProtocol.cpp
@@ -62,11 +62,11 @@ dataMove ServerProtocol::ReceiveAMove(bool& isConnected) {
     return newMove;
 }
 
-void ServerProtocol::SendMatch(const uint8_t& matchID, const uint8_t& quantityP,
-                               const uint8_t& maxP, bool& isConnected) {
+void ServerProtocol::SendMatch(const size_t& matchID, const uint8_t& quantityP, const uint8_t& maxP,
+                               bool& isConnected) {
     bool wasClosed = false;
     assistant.sendInt(A_MATCH, wasClosed);
-    assistant.sendInt(matchID, wasClosed);
+    assistant.sendInt(static_cast<uint8_t>(matchID), wasClosed);
     assistant.sendInt(quantityP, wasClosed);
     assistant.sendInt(maxP, wasClosed);
     if (wasClosed) {
